Move back stuff from model.py to loss.py
